feat: add option for additional entrypoints
- specify additonal entrypoints via plugin opt
- break up unit test file
- unit test additonal entrypoints
- test: add all tests to config
- test: use safe random strings for file and module names
- style: remove whitespace
