+++ ava.config.mjs
@@ -1,7 +1,7 @@
 const avaConfig = {
   extensions: ['ts'],
   failWithoutAssertions: false,
-  files: ['test/glob-plugin.test.ts'],
+  files: ['test/*.test.ts'],
   ignoredByWatcher: ['src/files/**/*'],
   require: ['ts-node/register'],
 };
