+++ test/glob-plugin.test.ts
@@ -1,40 +1,7 @@
-import untypedTest from 'ava';
-import del from 'del';
-import * as esbuild from 'esbuild';
 import { existsSync, promises as fs } from 'fs';
-import { customAlphabet } from 'nanoid';
-import path from 'path';
 
-import type { ExecutionContext, ImplementationFn, TestFn } from 'ava';
-
-import { globPlugin } from '../src';
-
-// CONFIG
-// ------
-
-// -- Directories used for test files
-const FILE_DIR_NAME = 'files';
-const IN_DIR_NAME = 'input';
-const OUT_DIR_NAME = 'output';
-const DEPENDENCY_DIR_NAME = 'dependencies';
-const FILE_DIR = path.resolve(__dirname, FILE_DIR_NAME);
-
-// -- Config for assertion retrying
-const MAX_RETRIES = 50;
-const RETRY_INTERVAL = 100;
-
-// -- Used to create directory and file names
-const nanoid = customAlphabet('ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz', 10);
-
-// -- Type the untyped
-interface TestContext {
-  /** Triggers a build */
-  build: () => void;
-  /** Unique directory for current test */
-  directory: string;
-}
-
-const test = untypedTest as TestFn<TestContext>;
+import { runner, test } from './runner';
+import { createEntryFile, retryAssertion, wait } from './util';
 
 // SINGLE RUN
 // ----------
@@ -180,239 +147,3 @@ test.serial(
     );
   }),
 );
-
-// UTILITY
-// -------
-
-/** Convenience function to create and write an entry file */
-async function createEntryFile({
-  directory,
-  withDependency = false,
-  withSyntaxError = false,
-}: {
-  directory: string;
-  withDependency?: boolean;
-  withSyntaxError?: boolean;
-}) {
-  const entryFile = new EntryFile({ directory });
-
-  if (withDependency) {
-    entryFile.addDependency();
-  }
-
-  if (withSyntaxError) {
-    entryFile.addSyntaxError();
-  }
-
-  await entryFile.write();
-  return entryFile;
-}
-
-/** Retries the assertion every x ms for a maximum amount of times */
-async function retryAssertion(
-  t: ExecutionContext,
-  assertion: ImplementationFn<any, any>,
-): Promise<void> {
-  return new Promise((resolve, reject) => {
-    let tried = 0;
-
-    function planNextCheck() {
-      setTimeout(async () => {
-        tried += 1;
-
-        const maxRetriesReached = tried >= MAX_RETRIES;
-        const attempt = await t.try(assertion);
-
-        if (attempt.passed) {
-          attempt.commit();
-          return resolve();
-        }
-
-        if (maxRetriesReached) {
-          attempt.commit();
-          return reject();
-        }
-
-        attempt.discard();
-        planNextCheck();
-      }, RETRY_INTERVAL);
-    }
-
-    planNextCheck();
-  });
-}
-
-async function wait(ms = 500) {
-  await new Promise((resolve) => setTimeout(resolve, ms));
-}
-
-// -- TEST RUNNER
-
-function runner(
-  testFunction: ImplementationFn<unknown[], TestContext>,
-  { watchMode = true } = {},
-): ImplementationFn<unknown[], TestContext> {
-  return async (t: ExecutionContext<TestContext>) => {
-    const directoryName = nanoid();
-    const directory = path.resolve(FILE_DIR, directoryName);
-
-    const inputDirectory = path.resolve(directory, IN_DIR_NAME);
-    const outputDirectory = path.resolve(directory, OUT_DIR_NAME);
-    const dependencyDirectory = path.resolve(directory, DEPENDENCY_DIR_NAME);
-
-    const [plugin, pluginControls] = globPlugin({
-      controls: true,
-    });
-
-    // Prepare
-    await Promise.all([
-      fs.mkdir(inputDirectory, { recursive: true }),
-      fs.mkdir(outputDirectory, { recursive: true }),
-      fs.mkdir(dependencyDirectory, { recursive: true }),
-    ]);
-
-    function build() {
-      const globPath = path.relative(process.cwd(), path.resolve(inputDirectory, '**/*'));
-
-      esbuild.build({
-        bundle: true,
-        entryPoints: [globPath.replace(/\\/g, '/')],
-        plugins: [plugin],
-        outdir: outputDirectory,
-        watch: watchMode,
-      });
-    }
-
-    if (watchMode) {
-      build();
-      await wait();
-    }
-
-    t.context.build = build;
-    t.context.directory = directory;
-
-    // Test
-    await (async () => {
-      await testFunction(t);
-    })().finally(async () => {
-      // Teardown
-      await pluginControls.stopWatching();
-      await del(directory);
-    });
-  };
-}
-
-// -- ENTRY FILE
-interface EntryFileRecipe {
-  name?: string;
-  directory: string;
-}
-
-class EntryFile {
-  public readonly directory: string;
-  public readonly name: string;
-  public readonly dependencies: Dependency[] = [];
-  public hasSyntaxError = false;
-
-  constructor({ name = nanoid(), directory }: EntryFileRecipe) {
-    this.name = name;
-    this.directory = directory;
-  }
-
-  public get contents() {
-    const contents: string[] = [];
-
-    // Imports
-    this.dependencies.forEach((dependency) => {
-      contents.push(dependency.importStatement);
-    });
-
-    // Content
-    contents.push(
-      `console.log('NAME', '${this.name}');`,
-      `console.log('RANDOM STRING', '${nanoid()}');`,
-      this.hasSyntaxError ? 'import;' : '',
-    );
-
-    // Dependency content
-    this.dependencies.forEach((dependency) => {
-      contents.push(dependency.codeBlock);
-    });
-
-    return contents.join('\n');
-  }
-
-  public get path() {
-    return path.resolve(this.directory, IN_DIR_NAME, `${this.name}.ts`);
-  }
-
-  public get outputPath() {
-    return path.resolve(this.directory, OUT_DIR_NAME, `${this.name}.js`);
-  }
-
-  public addDependency() {
-    this.dependencies.push(new Dependency({ directory: this.directory }));
-  }
-
-  public addSyntaxError() {
-    this.hasSyntaxError = true;
-  }
-
-  public removeSyntaxError() {
-    this.hasSyntaxError = false;
-  }
-
-  public async unlink({ dependencies = false, entry = true } = {}) {
-    if (dependencies) {
-      await Promise.all(this.dependencies.map((dependency) => fs.unlink(dependency.path)));
-    }
-
-    if (entry) {
-      await fs.unlink(this.path);
-    }
-  }
-
-  public async write({ dependencies = true, entry = true } = {}) {
-    if (dependencies) {
-      await Promise.all(
-        this.dependencies.map((dependency) => fs.writeFile(dependency.path, dependency.contents)),
-      );
-    }
-
-    if (entry) {
-      await fs.writeFile(this.path, this.contents);
-    }
-  }
-}
-
-// -- DEPENDENCY
-interface DependencyRecipe {
-  directory: string;
-  name?: string;
-}
-
-class Dependency {
-  public readonly name: string;
-  public readonly directory: string;
-
-  constructor({ directory, name = nanoid() }: DependencyRecipe) {
-    this.directory = directory;
-    this.name = name;
-  }
-
-  public get codeBlock() {
-    return `console.log('DEPENDENCY', ${this.name}());`;
-  }
-
-  public get contents() {
-    return `export const ${this.name} = () => '${this.name}';`;
-  }
-
-  public get importStatement() {
-    return `import { ${this.name} } from '../${DEPENDENCY_DIR_NAME}/${this.name}';`;
-  }
-
-  public get path() {
-    return path.resolve(this.directory, DEPENDENCY_DIR_NAME, `${this.name}.ts`);
-  }
-}
